Refactor API response stubs into a spec helper
We are using the same stubs over and over again in tests, moved these
into a test helper so they are reusable and tests are easier to read
