+++ spec/features/leasehold/letters/view_letter_preview_spec.rb
@@ -53,15 +53,6 @@ describe 'Viewing A Letter Preview' do
     expect(find('#successful_table')).to be_present
   end
 
-  def stub_my_cases_response
-    stub_const('Hackney::Income::IncomeApiUsersGateway', Hackney::Income::StubIncomeApiUsersGateway)
-
-    response_json = File.read(Rails.root.join('spec', 'examples', 'my_cases_response.json'))
-    stub_request(:get, /cases\?full_patch=false&is_paused=false&number_per_page=20&page_number=1&upcoming_court_dates=false&upcoming_evictions=false/)
-      .with(headers: { 'X-Api-Key' => ENV['INCOME_API_KEY'] })
-      .to_return(status: 200, body: response_json)
-  end
-
   def stub_get_templates_response
     stub_request(:get, %r{/messages\/letters\/get_templates})
       .with(headers: { 'X-Api-Key' => ENV['INCOME_API_KEY'] })
