+++ spec/rails_helper.rb
@@ -3,6 +3,7 @@ require 'spec_helper'
 require 'tenancy_helper'
 require 'document_helper'
 require 'support/mock_auth_helper'
+require 'support/api_response_helper'
 require 'support/chrome_test_helper'
 
 ENV['RAILS_ENV'] = 'test'
